backend/requirements.txt
functions-framework==3.*
flask==2.*
flask-cors==4.*
google-cloud-storage==2.*
google-cloud-pubsub==2.*
Werkzeug==2.*